add footer, first draf styling: footer and header
