+++ src/components/Footer.css
@@ -1,0 +1,15 @@
+body {
+    position: relative;
+    height: 100vh;
+    margin:0;
+  }
+
+.menu {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 50px;
+    display:flex;
+    justify-content: space-around;
+
+}
