+++ src/components/Login.css
@@ -1,10 +1,14 @@
+@import "https://fonts.googleapis.com/css2?family=Raleway:wght@700&display=swap";
+
 .login {
     margin: 0 auto;
-    width: 50%;
+    width: 50%;    font-family: 'Raleway', sans-serif;
+    font-weight: 500;
 }
 
-.login h1 {
-    text-align:center
+.login h3 {
+    text-align:center;
+    font-family: 'Raleway', sans-serif;
 }
 
 .login form {
