+++ src/components/Signup.css
@@ -1,10 +1,13 @@
+@import "https://fonts.googleapis.com/css2?family=Raleway:wght@700&display=swap";
+
 .signup {
     margin: 0 auto;
     width: 50%;
 }
 
-.signup h1 {
-    text-align:center
+.signup h3 {
+    text-align:center;
+    font-family: 'Raleway', sans-serif;
 }
 
 .signup form {
